LayoutTests/canvas/philip/tests/2d.shadow.image.transparent.2-expected.txt

< [index] >
2d.shadow.image.transparent.2
Shadows are not drawn for transparent parts of images
References: 2d.shadow.render
Actual output:
Expected output:

Passed
